core: widen state set ttl to i64 so an out-of-range value above u32::MAX is E_STATE_TTL; pin corpus to v1.0-rc.43 (AMB-27)
The parsed StateUpdateOp::Set ttl was typed u32, so a conforming 64-bit integer
above u32::MAX failed deserialization with a generic integer-width error mapped
to E_SCHEMA_FIELD_RANGE before the dedicated E_STATE_TTL hard-range check ran.
ttl (and STATE_TTL_HARD_RANGE) are widened to i64 (the section 04 integer
domain), so the dedicated E_STATE_TTL check governs any out-of-bounds value,
matching the Java reference and section 11:289. The min_refresh_interval
over-u32 case (a manifest field with no dedicated code) legitimately stays
E_SCHEMA_FIELD_RANGE.

Bump SPEC_REVISION 1.0-rc.42 -> 1.0-rc.43 and the CI corpus pin to ref
v1.0-rc.43 (corpus 83 vectors, including the new 169-state-ttl-over-u32).

+++ entangled-core/src/state/store.rs
@@ -289,7 +289,7 @@ impl StateStore {
             }
         }
 
-        let expires_at = *now + Duration::seconds(i64::from(ttl));
+        let expires_at = *now + Duration::seconds(ttl);
         let entry = StateEntry {
             value: value.clone(),
             mode,
